+++ aaem/components/hydro.py
@@ -156,13 +156,13 @@ def preprocess_existing_projects (ppo):
     names = []
     for p_idx in range(len(project_data)):
         cp = project_data.ix[p_idx]
-        p_name = 'hydro+' +str(cp['Project']).replace(' ','_').replace('/','-')
-        i = 1
-        i = 1
-        while (p_name in names):
-            p_name = 'hydro+' + str(cp['Project']).replace(' ','_').replace('/','-') + '_' + str(i)
-            i += 1
-        names.append(p_name)
+        p_name = 'hydro+project_' + str(p_idx) #+ str(cp['Project']).replace(' ','_').replace('/','-').replace(';','').replace('.','').replace("'",'').replace('_-_','_').replace(',','').replace('(','').replace(')','')
+        #~ i = 1
+        #~ i = 1
+        #~ while (p_name in names):
+            #~ p_name = 'hydro+' + str(cp['Project']).replace(' ','_').replace('/','-').replace(';','').replace('.','').replace("'",'').replace('_-_','_').replace(',','').replace('(','').replace(')','') + '_' + str(i)
+            #~ i += 1
+        #~ names.append(p_name)
 
         
         phase = cp['Phase Completed']
@@ -180,7 +180,8 @@ def preprocess_existing_projects (ppo):
         projects.append(p_name)
         
             
-        p_data[p_name] = {'phase': phase,
+        p_data[p_name] = {'name': str(cp['Project']),
+                    'phase': phase,
                     'proposed capacity': proposed_capacity,
                     'proposed generation': proposed_generation,
                     #~ 'distance to resource': distance_to_resource,
@@ -228,11 +229,14 @@ def component_summary (coms, res_dir):
             start_yr = hydro.comp_specs['start year']
             hydro.get_diesel_prices()
             diesel_price = float(hydro.diesel_prices[0].round(2))
+            #~ print hydro.diesel_prices[0]
             if not hydro.comp_specs["project details"] is None:
                 phase = hydro.comp_specs["project details"]['phase']
             else:
                 phase = "Reconnaissance"
-                
+
+            name = hydro.comp_specs["project details"]['name']
+            
             average_load = hydro.average_load
             proposed_load =  hydro.load_offset_proposed
             
@@ -268,6 +272,7 @@ def component_summary (coms, res_dir):
                 #~ red_per_year = 0
             
             l = [c, 
+                name,
                 start_yr,
                 phase,
 
@@ -291,12 +296,13 @@ def component_summary (coms, res_dir):
                 hydro.reason
             ]
             out.append(l)
-        except (KeyError,AttributeError) as e:
+        except (KeyError,AttributeError,TypeError) as e:
             #~ print e
             pass
         
     
     cols = ['Community',
+            'Project Name',
             'Start Year',
             'project phase',
             
@@ -311,10 +317,10 @@ def component_summary (coms, res_dir):
             'Hydro Power Reduction in Utility Diesel Consumed per year',
             'Diesel Generator Efficiency',
             'Diesel Price - year 1 [$]',
-            'Wind NPV benefits [$]',
-            'Wind NPV Costs [$]',
-            'Wind NPV Net benefit [$]',
-            'Wind Benefit Cost Ratio',
+            'Hydro NPV benefits [$]',
+            'Hydro NPV Costs [$]',
+            'Hydro NPV Net benefit [$]',
+            'Hydro Benefit Cost Ratio',
             'notes'
             ]
         
@@ -485,12 +491,9 @@ class Hydropower (AnnualSavings):
             self.net_generation_proposed = self.gross_generation_proposed -\
                                            tansmission_losses -\
                                            exess_energy
-                
-            start_year = self.comp_specs["start year"]
-            con = float(self.forecast.consumption.ix[start_year])
-            
-            if self.net_generation_proposed > con:
-                self.net_generation_proposed = con
+        
+            if self.net_generation_proposed > self.generation:
+                self.net_generation_proposed = self.generation
                                            
         #~ print 'self.load_offset_proposed', self.load_offset_proposed
         #~ print 'self.gross_generation_proposed', self.gross_generation_proposed
@@ -501,6 +504,7 @@ class Hydropower (AnnualSavings):
         """
         """
         # %
+       
         captured_percent = self.percent_excess_energy * \
                     self.comp_specs['percent excess energy capturable']
         
@@ -518,9 +522,17 @@ class Hydropower (AnnualSavings):
         gen_eff = self.cd["diesel generation efficiency"]
         self.generation_diesel_reduction = self.net_generation_proposed /\
                                             gen_eff
+                                            
+        #~ electric_diesel = self.generation /gen_eff
+        #~ if self.generation_diesel_reduction > electric_diesel:
+            #~ self.generation_diesel_reduction = electric_diesel
         
         # gal/year
-        self.lost_heat_recovery = self.generation_diesel_reduction * \
+        if not self.cd['heat recovery operational']:
+
+            self.lost_heat_recovery  = 0
+        else:
+            self.lost_heat_recovery = self.generation_diesel_reduction * \
                                 self.comp_specs['percent heat recovered']
         
         #~ print 'self.captured_energy', self.captured_energy
@@ -620,6 +632,7 @@ class Hydropower (AnnualSavings):
               "Hydro: Project Capital Cost ($/year)",
               "Hydro: Total Cost Savings ($/year)",
               "Hydro: Net Benefit ($/year)"]
+
         fname = os.path.join(directory,
                              self.cd['name'] + '_' + \
                              self.component_name + "_output.csv")